Update build scripts to use bin

+++ Scripts/install-release.sh
@@ -1,10 +1,19 @@
 #!/bin/sh
 
 if [ "$EUID" -ne 0 ]
-  then echo "Please run as root"
+  then echo -e "\e[33m Run as root!"
   exit
 fi
 
+cd ..
+
+FILE=bin/dynperf
+if ! test -f "$FILE"; then
+    echo -e "> Release binary not found"
+    echo -e "\e[33m Run ./build-release.sh first"
+    exit
+fi
+
 # Copy binary
-cd ./release
+cd ./bin
 cp dynperf /usr/bin/dynperf
